refactor: 型の置き場所の整理
