+++ src/features/template-search/store/types.ts
@@ -5,3 +5,15 @@ export type FavoriteTemplateList = {
 export type TemplateSearchSettings = {
     featureUsage: boolean
 }
+
+export type Template = DOMElement & {
+    isFavorite: boolean,
+    id: string
+}
+
+export type DOMElement = {
+    title: string,
+    href: string,
+    index: number,
+    rawElement: Element,
+}
